pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "planner-ai-platform"
version = "0.1.0"
description = "Planner AI platform: schema, validator, linter, expansion engine"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Patrick"}]
dependencies = [
  "typer>=0.12.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff",
]
agents = [
  "openai>=1.40.0",
  "rich>=13.7.0",
]

[project.scripts]
planner = "planner_ai_platform.cli:cli"
planner-agent = "planner_agent.cli:cli"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["planner_ai_platform*", "planner_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Keep it light: Ruff defaults are already conservative.
# (If you explicitly want: select = ["E4","E7","E9","F"])